Switch some mappers over to splitbits!

+++ src/memory/mappers/mapper001_0.rs
@@ -77,9 +77,9 @@ impl Mapper for Mapper001_0 {
                 0x6000..=0x7FFF => unreachable!(),
                 0x8000..=0x9FFF => {
                     let fields = splitbits!(finished_value, "...cppmm");
-                    params.set_chr_layout(CHR_LAYOUTS[fields.c]);
-                    params.set_prg_layout(PRG_LAYOUTS[fields.p]);
-                    params.set_name_table_mirroring(MIRRORINGS[fields.m]);
+                    params.set_chr_layout(CHR_LAYOUTS[fields.c as usize]);
+                    params.set_prg_layout(PRG_LAYOUTS[fields.p as usize]);
+                    params.set_name_table_mirroring(MIRRORINGS[fields.m as usize]);
                 }
                 // FIXME: Handle cases for special boards.
                 0xA000..=0xBFFF => params.set_bank_register(C0, finished_value),
@@ -87,8 +87,8 @@ impl Mapper for Mapper001_0 {
                 0xC000..=0xDFFF => params.set_bank_register(C1, finished_value),
                 0xE000..=0xFFFF => {
                     let fields = splitbits!(finished_value, "...rbbbb");
-                    params.set_ram_status(S0, RAM_STATUSES[fields.r]);
-                    params.set_bank_register(P0, fields.b as u16);
+                    params.set_ram_status(S0, RAM_STATUSES[fields.r as usize]);
+                    params.set_bank_register(P0, fields.b);
                 }
             }
         }
